Implemented Financial Report Generator (#86)
* Feature: Financial Report Generator
Generates a pdf report for the current user. The pdf is displayed in a new tab once generated. Contains all transactions for a user.

* Feature: Generate PDF Summary
Addition of new models to assist in generation of summary

* Chore: gitignore generated summary pdf

* Fix: Summary does not account of incomes in the present month
Changed to account for all incomes and expenses from the current month. Any transactions from future months are left out.